move type defs to InvariantEncoderSpec
